bug: fixup multiline comment handling in import spec group doc comments
Fix a bug where Import Group Doc multiline comments in the form of
`/* <lines> */` led to a missing new line, on the first ImportSpec in the
group.

This is remediated by asserting a new line at the end of each Doc comment in the builder.

Example:

```go
import (
  /*
    Multiline comment
  */
  "spec"
)
```

Rendered as:

```go
import (
  /*
    Multiline comment
  */"spec"
)
```